+++ api4jenkins/__init__.py
@@ -10,7 +10,7 @@ from .__version__ import (__author__, __author_email__, __copyright__,
                           __description__, __license__, __title__, __url__,
                           __version__)
 from .credential import Credentials
-from .exceptions import ItemNotFoundError
+from .exceptions import ItemNotFoundError, AuthenticationError
 from .item import Item
 from .job import Folder
 from .node import Nodes
@@ -73,7 +73,7 @@ class Jenkins(Item):
             >>> print(job)
             <FreeStyleProject: http://127.0.0.1:8080/job/freestylejob/>
         '''
-        folder, name = self._get_folder(full_name)
+        folder, name = self._resolve_name(full_name)
         return folder.get(name)
 
     def iter_jobs(self, depth=0):
@@ -117,7 +117,7 @@ class Jenkins(Item):
             >>> print(job)
             <FreeStyleProject: http://127.0.0.1:8080/job/freestylejob/>
         '''
-        folder, name = self._get_folder(full_name)
+        folder, name = self._resolve_name(full_name)
         return folder.create(name, xml)
 
     def copy_job(self, full_name, dest):
@@ -136,7 +136,7 @@ class Jenkins(Item):
             >>> print(job)
             <FreeStyleProject: http://127.0.0.1:8080/job/folder/job/newjob/>
         '''
-        folder, name = self._get_folder(full_name)
+        folder, name = self._resolve_name(full_name)
         return folder.copy(name, dest)
 
     def delete_job(self, full_name):
@@ -187,6 +187,10 @@ class Jenkins(Item):
             raise ItemNotFoundError(f'No such job: {full_name}')
         return job.build(**params)
 
+    def check_job_name(self, name):
+        resp = self.handle_req('GET', 'checkJobName', params={'value': name})
+        return 'is an unsafe character' in resp.text
+
     def _url2name(self, url):
         '''Covert job url to full name
 
@@ -208,8 +212,8 @@ class Jenkins(Item):
         full_name = full_name.strip('/').replace('/', '/job/')
         return f'{self.url}job/{full_name}/'
 
-    def _get_folder(self, full_name):
-        '''Split folder and job'''
+    def _resolve_name(self, full_name):
+        '''Resolve folder and job name from full name'''
         path = PurePosixPath(full_name)
         parent = str(path.parent) if path.parent.name else ''
         return Folder(self, self._name2url(parent)), path.name
@@ -232,7 +236,16 @@ class Jenkins(Item):
     @property
     def crumb(self):
         '''Crumb of Jenkins'''
-        self._add_crumb({})
+        if self._crumb is None:
+            try:
+                _crumb = self.send_req(
+                    'GET', self.url + 'crumbIssuer/api/json').json()
+                self._crumb = {_crumb['crumbRequestField']: _crumb['crumb']}
+            except HTTPError as e:
+                if e.response.status_code in [401, 403]:
+                    raise AuthenticationError(
+                        'Invalid authorization for %s' % self) from e
+                self._crumb = {}
         return self._crumb
 
     @property
@@ -284,6 +297,15 @@ class Jenkins(Item):
     @property
     def me(self):
         return self.user
+
+    def __iter__(self):
+        yield from self.iter_jobs()
+
+    def __call__(self, depth):
+        yield from self.iter_jobs(depth)
+
+    def __getitem__(self, full_name):
+        return self.get_job(full_name)
 
 
 def _patch_to(module, cls, func=None):